Extract automation inline handlers into yaml/inline.py
Move ``upsert_inline_handler`` / ``remove_inline_handler`` and their
helpers (``_locate_component_instance``, ``_instance_id_matches``,
``_indent_block``) out of ``helpers/yaml/__init__.py`` into a new
``helpers/yaml/inline.py`` submodule. ``__init__.py`` re-exports the
public callables (and ``_indent_block``, which tests import directly)
so existing ``from .helpers.yaml import upsert_inline_handler`` callers
keep working unchanged.

Byte-for-byte structural move; no behaviour change. Final PR in the
helpers/yaml split arc — ``__init__.py`` is now a 57-line re-export
shell with the libyaml loader pick and the package-level re-export
list. Every submodule sits well under the 800-line soft cap.
